requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
sqlalchemy==1.4.23
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.5
pandas==1.3.3
openpyxl==3.0.9
python-dotenv==0.19.0
mysql-connector-python==8.0.26
bcrypt==4.0.1
numpy==1.21.0
